mon: Clear migration_src from target pool when pool migration finishes
We want to clear the migration_src value from the target pool when a pool
migration finishes. This means the pool will no longer be considered to be
part of an active migration, and so can be used as the source pool in a new migration.
We do not want to clear the migration_target value from the source pool as
we want this to persist in order to redirect traffic from the source to the
target.

Signed-off-by: Jamie Pryde <jamiepry@uk.ibm.com>