osd: pool migration schedule deletion of objects
The last step of migrating an object to another pool will be to delete
it from the source pool. This commit makes recover_pool_migration
put the appropriate locks in place (write object rwlock and also block
future writes) and schedule the deletion of the next object to be
migrated. When this completes queue_recovery is called to continue
the migration process. Previous commits update the migration watermark
as objects are deleted, so the result is that a pool migration now
schedules the delete of all objects in the source pool.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>